Create first unit test

+++ src/main/java/dev/nichoko/diogenes/model/dto/ItemDTO.java
@@ -1,24 +1,24 @@
 package dev.nichoko.diogenes.model.dto;
 
 public class ItemDTO {
-    private Long id;
+    private int id;
     private String name;
     private String description;
-    private Long number;
+    private int number;
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
 
-    public Long getId() {
+    public int getId() {
         return id;
     }
 
-    public Long getNumber() {
+    public int getNumber() {
         return number;
     }
 
-    public void setNumber(Long number) {
+    public void setNumber(int number) {
         this.number = number;
     }
